+++ src/subapps/bizlinksfree/bizlinksfree.controller.ts
@@ -11,8 +11,8 @@ unnecessary, BUT eventually will need editing.
 */
 
 import { Controller, Get } from '@nestjs/common';
-import { BizlinksfreeUserSettingsService } from './services/bizlinksfreeUserSettings.service';
-import { BizlinksfreeUrlService } from './services/bizlinksfreeUrl.service';
+import { BizlinksfreeUserSettingsService } from './resources/services/bizlinksfreeUserSettings.service';
+import { BizlinksfreeUrlService } from './resources/services/bizlinksfreeUrl.service';
 // import { CreateBizlinksfreeDto } from './dto/create-bizlinksfree.dto';
 // import { UpdateBizlinksfreeDto } from './dto/update-bizlinksfree.dto';
 import { Auth } from 'src/iam/decorators/auth.decorator';
