+++ src/subapps/bizlinksfree/resources/services/bizlinksfreeUserSettings.service.ts
@@ -13,8 +13,8 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { BizlinksfreeUserSettings } from '../entities/bizlinksfreeUserSettings.entity';
-import { CreateBizlinksfreeUserSettingsDto } from '../dto/create-bizlinksfree-user-settings.dto';
+import { BizlinksfreeUserSettings } from '../../entities/bizlinksfreeUserSettings.entity';
+import { CreateBizlinksfreeUserSettingsDto } from '../../dto/create-bizlinksfree-user-settings.dto';
 // import { UpdateBizlinksfreeUserSettingsDto } from './dto/update-bizlinksfree-user-settings.dto';
 // import { CreateBizlinksfreeUrlDto } from './dto/create-bizlinksfree-url.dto';
 // import { UpdateBizlinksfreeUrlDto } from './dto/update-bizlinksfree-url.dto';
