+++ src/subapps/mycuttingboard/mycuttingboard.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 
-import { BoardsService } from './resources/controllers/services/boards.service';
-import { LinksService } from './resources/controllers/services/links.service';
-import { WoodsService } from './resources/controllers/services/woods.service';
+import { BoardsService } from './resources/services/boards.service';
+import { LinksService } from './resources/services/links.service';
+import { WoodsService } from './resources/services/woods.service';
 
 // import { CreateBoardDto } from './dto/create-board.dto';
 // import { UpdateBoardDto } from './dto/update-board.dto';
