+++ src/subapps/mycuttingboard/resources/services/links.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { MycuttingboardLinks } from '../../../entities/mycuttingboardLinks.entity';
+import { MycuttingboardLinks } from '../../entities/mycuttingboardLinks.entity';
 @Injectable()
 export class LinksService {
   constructor(
